+++ test/integration/service-mixins.spec.js
@@ -127,6 +127,9 @@ describe("Test Service mixins", () => {
 					msg: "From mixin1L1",
 					action: _.omit(ctx.action, ["handler", "service", "metrics"])
 				};
+			},
+			foxtrot() {
+				return "From mixin1L1";
 			}
 		},
 
@@ -263,7 +266,9 @@ describe("Test Service mixins", () => {
 			},
 			zulu() {
 				throw new MoleculerError("Zulu error");
-			}
+			},
+
+			foxtrot: false
 		},
 
 		methods: {
@@ -464,6 +469,12 @@ describe("Test Service mixins", () => {
 				"mixin2L1-error-all-1",
 				"mixin2L1-error-all-2",
 			]);
+		});
+	});
+
+	it("should not call 'foxtrot' action", () => {
+		return broker.call("main.foxtrot").catch(err => {
+			expect(err.name).toBe("ServiceNotFoundError");
 		});
 	});